+++ docs/conf.py
@@ -17,7 +17,7 @@ version = "0.1.0"
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx_copybutton", "sphinx_tabs.tabs"]
+extensions = ["sphinx.ext.autodoc", "sphinx_copybutton", "sphinx_tabs.tabs"]
 
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]