+++ pdfino/type_definitions.py
@@ -10,9 +10,8 @@ from reportlab.lib.styles import LineStyle, ListStyle, ParagraphStyle
 class Pagesize(NamedTuple):
     """NamedTuple representing a page size in points, for a resolution of 72 dpi.
 
-    Args:
-        width: Width of the page.
-        height: Height of the page.
+    :param width: Width of the page in points.
+    :param height: Height of the page in points.
     """
 
     width: float
@@ -22,14 +21,9 @@ class Pagesize(NamedTuple):
     def from_name(cls, pagesize: str) -> "Pagesize":
         """Alternative constructor to create a Pagesize from a string, e.g. "A4".
 
-        Args:
-            pagesize: ReportLab pagesize.
-
-        Returns:
-            Pagesize named tuple.
-
-        Raises:
-            ValueError: If the pagesize is not a valid constant of `reportlab.lib.pagesizes`.
+        :param pagesize: Name of the pagesize, e.g. "A4".
+        :return: Pagesize named tuple.
+        :raises ValueError: If the pagesize is not a valid constant of `reportlab.lib.pagesizes`.
         """
         try:
             return cls(*getattr(pagesizes, pagesize.upper()))
@@ -40,11 +34,10 @@ class Pagesize(NamedTuple):
 class Margins(NamedTuple):
     """NamedTuple representing margins.
 
-    Args:
-        top: Top margin.
-        right: Right margin.
-        bottom: Bottom margin.
-        left: Left margin.
+    :param top: Top margin.
+    :param right: Right margin.
+    :param bottom: Bottom margin.
+    :param left: Left margin.
     """
 
     top: int
@@ -56,13 +49,12 @@ class Margins(NamedTuple):
 class Font(NamedTuple):
     """NamedTuple representing a font definition with different variants.
 
-    Args:
-        name: Name of the font.
-        normal: Path to the normal variant of the font.
-        bold: Path to the bold variant of the font.
-        italic: Path to the italic variant of the font.
-        bold_italic: Path to the bold italic variant of the font.
-        default: Whether this font should be used as the default font for a Template.
+    :param name: Name of the font.
+    :param normal: Path to the normal variant of the font.
+    :param bold: Path to the bold variant of the font.
+    :param italic: Path to the italic variant of the font.
+    :param bold_italic: Path to the bold italic variant of the font.
+    :param default: Whether this font should be used as the default font for a Template.
     """
 
     name: str
@@ -74,13 +66,23 @@ class Font(NamedTuple):
 
 
 class StyleOptions(TypedDict, total=False):
-    """Styling options for an element."""
+    """Styling options for an element.
+
+    :param color: The color of the element.
+    """
 
     color: str
 
 
 class LayoutOptions(TypedDict, total=False):
-    """Layout options for a page."""
+    """Layout options for a page.
+
+    :param margin_top: The top margin of the page.
+    :param margin_right: The right margin of the page.
+    :param margin_bottom: The bottom margin of the page.
+    :param margin_left: The left margin of the page.
+    :param margins: The margins of the page.
+    """
 
     margin_top: int
     margin_right: int
@@ -90,7 +92,10 @@ class LayoutOptions(TypedDict, total=False):
 
 
 class ElementOptions(LayoutOptions, StyleOptions, total=False):
-    """All options for an element."""
+    """All options for an element.
+
+    :param align: The alignment of the element.
+    """
 
     align: Literal["left", "right", "center", "justify"]
 
@@ -98,13 +103,14 @@ class ElementOptions(LayoutOptions, StyleOptions, total=False):
 class Style(NamedTuple):
     """NamedTuple representing a style definition.
 
-    Args:
-        name: Name of the style.
-        parent: Parent style to inherit from (name of another style)
-        options: Options for the style.
-        font_name: Name of the font to use.
-        font_size: Size of the font.
-        line_height: Line height of the style, relative to the font size.
+    :param name: Name of the style.
+    :param parent: Parent style to inherit from (name of another style)
+    :param options: Options for the style.
+    :param font_name: Name of the font to use.
+    :param font_size: Size of the font.
+    :param line_height: Line height of the style, relative to the font size.
+    :param allow_widows: Whether to allow widows (single lines at the end of a paragraph).
+    :param hyphenate: Whether to hyphenate the text (requires `pyphen` to be installed)
     """
 
     name: str
@@ -113,6 +119,8 @@ class Style(NamedTuple):
     font_name: Optional[str] = None
     font_size: Optional[Union[int, float]] = None
     line_height: Optional[Union[int, float]] = None
+    allow_widows: bool = False
+    hyphenate: bool = False
 
 
 ReportLabStyle = Union[ParagraphStyle, ListStyle, LineStyle]